Refactor module file
