+++ modules/atomic-opt-in/assets/js/opt-in-page/opt-in-img-placeholders.js
