+++ modules/atomic-opt-in/assets/js/welcome-screen/app.js
