+++ modules/atomic-opt-in/assets/js/welcome-screen/welcome-popover.js
